requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
pillow
requests
sentencepiece
protobuf
accelerate
open-clip-torch
huggingface_hub